Update cryptopp source to version 8.7.0

+++ src/validat0.cpp
@@ -1575,7 +1575,10 @@ bool TestASN1Functions()
 
     {
         const std::string message = "Now is the time for all good men to come to the aide of their country";
-        const int asnStringTypes[] = {UTF8_STRING, PRINTABLE_STRING, T61_STRING, VIDEOTEXT_STRING, IA5_STRING, VISIBLE_STRING};
+        const byte asnStringTypes[] = {
+            UTF8_STRING, PRINTABLE_STRING, T61_STRING, VIDEOTEXT_STRING,IA5_STRING, VISIBLE_STRING
+        };
+
         unsigned int failed = 0;
         size_t len = 0, rlen = 0, i = 0;
 
@@ -1606,7 +1609,7 @@ bool TestASN1Functions()
     {
         const byte date[] = "Sun, 21 Mar 2021 01:00:00 +0000";
         SecByteBlock message; message.Assign(date, sizeof(date)-1);
-        const int asnDateTypes[] = {UTC_TIME, GENERALIZED_TIME};
+        const byte asnDateTypes[] = {UTC_TIME, GENERALIZED_TIME};
         unsigned int failed = 0;
         size_t i = 0;
